+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 11.0.2-20240911-1
+
+* Change options to store bool values as strings.
+* Update the quest log tooltip to have WQ related info.
+
 ## 11.0.2-20240910-2
 
 * Added workarounds for World Map taints.